backend/requirements.txt
# FastAPI Backend Requirements for REMHART Digital Twin
# Python 3.10 or higher

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# CORS & Environment
python-dotenv==1.0.0

# Data Processing & ML (for future AI models)
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2

# Date/Time handling
python-dateutil==2.8.2

# HTTP Client (for testing)
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
